+++ src/Explosion.cpp
@@ -4,7 +4,7 @@
 
 #include "Explosion.hpp"
 
-Bomberman::Explosion::Explosion(int x, int y) : _irr(Bomberman::Irrlicht::instance()) {
+Bomberman::Explosion::Explosion(float x, float y) : _irr(Bomberman::Irrlicht::instance()) {
   this->_ps = this->_irr.getSmgr()->addParticleSystemSceneNode(false);
   this->_em = _ps->createBoxEmitter(irr::core::aabbox3d<irr::f32>(-7,0,-7,7,1,7),
 				    irr::core::vector3df(0.0f,0.0f,0.0f), 10, 50,
